requirements.txt
Flask
rembg
Pillow
opencv-python
flask-cors
